refactor : Calculator 클래스를 domain에서 util로 이동
