+++ src/main/java/util/Calculator.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
